IMP: Конфигурация, соединение с СУБД и утилиты теперь в отдельном модуле
